+++ contracts/interfaces/IVotingEscrow.sol
@@ -7,14 +7,19 @@ interface IVotingEscrow {
     struct Point {
         int128 bias;
         int128 slope;
-        uint256 ts;
-        uint256 blk; // TODO get rid of blk?
+        uint ts;
+        uint blk; // TODO get rid of blk?
     }
 
-    function epoch() external view returns(uint256);
-    function user_point_epoch(address _user) external returns (uint256);
-    function user_point_history(address _user, uint256 _epoch) external returns (Point memory);
-    function point_history(uint256 _epoch) external returns (Point memory);
+    function epoch() external view returns (uint);
+
+    function user_point_epoch(address _user) external returns (uint);
+
+    function user_point_history(address _user, uint _epoch) external returns (Point memory);
+
+    function point_history(uint _epoch) external returns (Point memory);
+
     function checkpoint() external;
-    function locked__end(address user) external view returns (uint256);
+
+    function locked__end(address user) external view returns (uint);
 }
